+++ consul/state/state_store.go
@@ -1248,7 +1248,13 @@ func (s *StateStore) parseCheckServiceNodes(
 		return 0, nil, err
 	}
 
-	var results structs.CheckServiceNodes
+	// Special-case the zero return value to nil, since this ends up in
+	// external APIs.
+	if len(services) == 0 {
+		return idx, nil, nil
+	}
+
+	results := make(structs.CheckServiceNodes, 0, len(services))
 	for _, sn := range services {
 		// Retrieve the node.
 		n, err := tx.First("nodes", "id", sn.Node)