+++ CHANGELOG.md
@@ -32,6 +32,10 @@ FEATURES:
   quorum. This version also provides a foundation for new features that will
   appear in future Consul versions once the remainder of the v2 library is
   complete. [GH-2222]
+* Added new `consul operator` command, HTTP endpoint, and associated ACL to
+  allow Consul operators to view and update the Raft configuration. This allows
+  for a stale server to be removed without requiring downtime and peers.json
+  recovery file use. [GH-2312]
 * Extended the [`translate_wan_addrs`](https://www.consul.io/docs/agent/options.html#translate_wan_addrs)
   config option to also translate node addresses in HTTP responses, making it easy
   to use this feature from non-DNS clients. [GH-2118]
@@ -54,6 +58,9 @@ BACKWARDS INCOMPATIBILITIES:
 * `skip_leave_on_interrupt`'s default behavior is now dependent on whether or
   not the agent is acting as a server or client.  When Consul is started as a
   server the default is `true` and `false` when a client. [GH-1909]
+* `allow_stale` for DNS queries now defaults to `true`, allowing for better
+  utilization of available Consul servers and higher throughput at the exponse of
+  weaker consistency. [GH-2315]
 * HTTP check output is truncated to 4k, similar to script check output. [GH-1952]
 * Consul's Go API client will now send ACL tokens using HTTP headers instead of
   query parameters, requiring Consul 0.6.0 or later. [GH-2233]